+++ controls/models.py
@@ -10,7 +10,7 @@ from guardian.shortcuts import (assign_perm, get_objects_for_user,
 from simple_history.models import HistoricalRecords
 from jsonfield import JSONField
 from natsort import natsorted
-from .oscal import Catalogs, Catalog
+from controls.oscal import Catalogs, Catalog
 import uuid
 import tools.diff_match_patch.python3 as dmp_module
 from copy import deepcopy
@@ -257,22 +257,58 @@ class Element(auto_prefetch.Model):
         except:
             return False
 
+    @transaction.atomic
+    def remove_element_control(self, oscal_ctl_id, oscal_catalog_key):
+        """Remove a selected control from a system.root_element"""
+
+        try:
+            if ElementControl.objects.filter(element=self, oscal_ctl_id=oscal_ctl_id, oscal_catalog_key=oscal_catalog_key).exists():
+                ElementControl.objects.get(element=self, oscal_ctl_id=oscal_ctl_id, oscal_catalog_key=oscal_catalog_key).delete()
+            result = True
+        except:
+            result = False
+
+        return result
+
+    @transaction.atomic
     def assign_baseline_controls(self, user, baselines_key, baseline_name):
-        """Assign set of controls from baseline to an element"""
+        """Assign set of controls from baseline to system.root_element"""
 
         # Usage
             # s = System.objects.get(pk=20)
-            # s.root_element.assign_baseline_controls('NIST_SP-800-53_rev4', 'low')
+            # s.root_element.assign_baseline_controls(user, 's', 'low')
 
-        can_assign_controls = user.has_perm('change_element', self)
+        # Get system's existing selected controls and build list control ids
+        selected_controls_cur = self.controls.all()
+        selected_controls_ids_cur = set([f"{sc.oscal_ctl_id}=+={sc.oscal_catalog_key}" for sc in selected_controls_cur])
+
+        # Create object to track controls added, removed, and no_change in existing selected controls
+        changed_controls = {"add": [], "remove": [], "no_change": []}
+
         # Does user have edit permissions on system?
-        if  can_assign_controls:
-            from controls.models import Baselines
+        can_assign_controls = user.has_perm('change_element', self)
+        if can_assign_controls:
             bs = Baselines()
             controls = bs.get_baseline_controls(baselines_key, baseline_name)
             for oscal_ctl_id in controls:
-                ec = ElementControl(element=self, oscal_ctl_id=oscal_ctl_id, oscal_catalog_key=baselines_key)
-                ec.save()
+                if f"{oscal_ctl_id}=+={baselines_key}" in selected_controls_ids_cur:
+                    # Control already in selected, just append to 'no_change' list
+                    changed_controls['no_change'].append(f"{oscal_ctl_id}=+={baselines_key}")
+                    next
+                else:
+                    # Control in in selected, add control to selected controls and append to 'add' list
+                    ec = ElementControl(element=self, oscal_ctl_id=oscal_ctl_id, oscal_catalog_key=baselines_key)
+                    ec.save()
+                    changed_controls['add'].append(f"{oscal_ctl_id}=+={baselines_key}")
+            # We are done adding new controls to selected
+            # Now remove controls previously selected but not in new baseline
+            selected_controls_ids_new = set([f"{oscal_ctl_id}=+={baselines_key}" for oscal_ctl_id in controls])
+            for scc in selected_controls_ids_cur:
+                if scc not in selected_controls_ids_new:
+                    oscal_ctl_id_rm = scc.split("=+=")[0]
+                    remove_result = self.remove_element_control(oscal_ctl_id_rm, baselines_key)
+                    if remove_result:
+                        changed_controls['remove'].append(scc)
             return True
         else:
             # print("User does not have permission to assign selected controls to element's system.")