+++ CHANGELOG.md
@@ -4,6 +4,10 @@ GovReady-Q Release Notes
 v999 (March XX, 2021)
 ----------------------------
 
+**Feature change**
+
+* Re-assign system's baseline to different baseline; dynamically batch add and removes controls to change a system's existing baseline to a different baseline (e.g. from moderate to low)
+
 
 v0.9.3.0rc1 (March 16, 2021)
 ----------------------------